Move debugger classes
