+++ src/BlazeServiceProvider.php
@@ -5,7 +5,9 @@ namespace Livewire\Blaze;
 use Livewire\Blaze\Commands\TraceClearCommand;
 use Livewire\Blaze\Commands\TraceListCommand;
 use Livewire\Blaze\Commands\TraceShowCommand;
-use Livewire\Blaze\Compiler\Profiler;
+use Livewire\Blaze\Debugger\Debugger;
+use Livewire\Blaze\Debugger\DebuggerMiddleware;
+use Livewire\Blaze\Debugger\Instrumenter;
 use Livewire\Blaze\Runtime\BlazeRuntime;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Blade;
@@ -23,7 +25,7 @@ class BlazeServiceProvider extends ServiceProvider
         $this->app->singleton(BlazeRuntime::class);
         $this->app->singleton(Config::class);
         $this->app->singleton(Debugger::class);
-        $this->app->singleton(Profiler::class);
+        $this->app->singleton(Instrumenter::class);
         $this->app->singleton(BlazeManager::class);
 
         $this->app->singleton(\PhpParser\Parser::class, function () {
